// SmartBalance/Version 1.0.0/Code/Frontend/mersal-accounting-app/src/app/BusinessFeatures/Document/Models/search.model.ts
// export class Search {      
//     bank: {value:any};
//     journalType: {value:any};
//     DateTo: Date;
//     DateFrom: Date;  
// }
export class PostSearch {
    id: any;
    
    documentNumber: any;

    countReceipts: number

    firstNumber: any;

    pageIndex: number;
    pageSize: number;
    filters: any[];
    sort: any[];
}